Create Send message
This VBA code will open the whatsapp desktop app and will send multiple messages (as many as the workbook can handle). It requires a phone number and a message. It also can send images but after the whatsapp update it will break and will send about 50% of the images and will have different issues if also sent with text.