Create Template creator
This VBA macro will help create multiple templates depending on the variant that you choose. 
You will have to choose a file to load the variants on the macro. 
Create parameters to create templates. e.g. home state, age, or any parameter that is available on the first file. 
And last, uploading the file with all the information so it can be categorized and create a file that will provide the personalized message with the phone number to just start sending the messages.